Unify attachment handling
