+++ app/lib/features/attachments/types.dart
@@ -1,0 +1,8 @@
+import 'dart:io';
+
+import 'package:acter/common/models/types.dart';
+
+typedef OnAttachmentSelected = Future<void> Function(
+  List<File> files,
+  AttachmentType attachmentType,
+);
