+++ app/lib/features/attachments/widgets/attachment_section.dart
@@ -1,16 +1,22 @@
-import 'package:acter/common/dialogs/attachment_selection.dart';
+import 'dart:io';
+import 'dart:typed_data';
+
+import 'package:acter/common/models/types.dart';
 import 'package:acter/common/toolkit/buttons/danger_action_button.dart';
 import 'package:acter/common/widgets/input_text_field.dart';
+import 'package:acter/features/attachments/actions/select_attachment.dart';
 import 'package:acter/features/attachments/providers/attachment_providers.dart';
 import 'package:acter/features/attachments/widgets/attachment_item.dart';
+import 'package:acter/features/home/providers/client_providers.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart'
-    show Attachment, AttachmentsManager;
+    show Attachment, AttachmentDraft, AttachmentsManager;
 import 'package:atlas_icons/atlas_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:logging/logging.dart';
+import 'package:mime/mime.dart';
 import 'package:skeletonizer/skeletonizer.dart';
 
 final _log = Logger('a3::common::attachments');
@@ -99,7 +105,7 @@ class FoundAttachmentSectionWidget extends ConsumerWidget {
                   if (list.isNotEmpty)
                     for (var item in list)
                       _buildAttachmentItem(context, item, canEdit),
-                  if (canEdit) _buildAddAttachment(context, attachmentManager),
+                  if (canEdit) _buildAddAttachment(context, ref),
                 ],
               ),
             ],
@@ -249,13 +255,17 @@ class FoundAttachmentSectionWidget extends ConsumerWidget {
     }
   }
 
-  Widget _buildAddAttachment(BuildContext context, AttachmentsManager manager) {
+  Widget _buildAddAttachment(BuildContext context, WidgetRef ref) {
     final containerColor = Theme.of(context).colorScheme.surface;
     final iconColor = Theme.of(context).colorScheme.secondary;
     final iconTextStyle = Theme.of(context).textTheme.labelLarge;
     return InkWell(
       key: AttachmentSectionWidget.addAttachmentBtnKey,
-      onTap: () => showAttachmentSelection(context, manager),
+      onTap: () => selectAttachment(
+        context: context,
+        onSelected: (files, selectedType) =>
+            handleAttachmentSelected(context, ref, files, selectedType),
+      ),
       child: Container(
         height: 100,
         width: 100,
@@ -275,5 +285,75 @@ class FoundAttachmentSectionWidget extends ConsumerWidget {
         ),
       ),
     );
+  }
+
+  // if generic attachment, send via manager
+  Future<void> handleAttachmentSelected(
+    BuildContext context,
+    WidgetRef ref,
+    List<File> attachments,
+    AttachmentType attachmentType,
+  ) async {
+    /// converts user selected media to attachment draft and sends state list.
+    /// only supports image/video/audio/file.
+    final lang = L10n.of(context);
+    EasyLoading.show(status: lang.sendingAttachment);
+    final client = ref.read(alwaysClientProvider);
+    final manager = attachmentManager;
+    List<AttachmentDraft> drafts = [];
+    try {
+      for (var selected in attachments) {
+        final file = selected;
+        final mimeType = lookupMimeType(file.path);
+        if (mimeType == null) throw lang.failedToDetectMimeType;
+        if (attachmentType == AttachmentType.camera ||
+            attachmentType == AttachmentType.image) {
+          Uint8List bytes = await file.readAsBytes();
+          final decodedImage = await decodeImageFromList(bytes);
+          final imageDraft = client
+              .imageDraft(file.path, mimeType)
+              .size(bytes.length)
+              .width(decodedImage.width)
+              .height(decodedImage.height);
+          final attachmentDraft = await manager.contentDraft(imageDraft);
+          drafts.add(attachmentDraft);
+        } else if (attachmentType == AttachmentType.audio) {
+          Uint8List bytes = await file.readAsBytes();
+          final audioDraft =
+              client.audioDraft(file.path, mimeType).size(bytes.length);
+          final attachmentDraft = await manager.contentDraft(audioDraft);
+          drafts.add(attachmentDraft);
+        } else if (attachmentType == AttachmentType.video) {
+          Uint8List bytes = await file.readAsBytes();
+          final videoDraft =
+              client.videoDraft(file.path, mimeType).size(bytes.length);
+          final attachmentDraft = await manager.contentDraft(videoDraft);
+          drafts.add(attachmentDraft);
+        } else {
+          String fileName = file.path.split('/').last;
+          final fileDraft = client
+              .fileDraft(file.path, mimeType)
+              .filename(fileName)
+              .size(file.lengthSync());
+          final attachmentDraft = await manager.contentDraft(fileDraft);
+          drafts.add(attachmentDraft);
+        }
+      }
+      for (var draft in drafts) {
+        final res = await draft.send();
+        _log.info('attachment sent: $res');
+      }
+      EasyLoading.dismiss();
+    } catch (e) {
+      _log.severe('Error sending attachments', e);
+      if (!context.mounted) {
+        EasyLoading.dismiss();
+        return;
+      }
+      EasyLoading.showError(
+        lang.errorSendingAttachment(e),
+        duration: const Duration(seconds: 3),
+      );
+    }
   }
 }
